Gogeo math32 pkg integration

+++ src/server/actions/actions.go
@@ -47,7 +47,7 @@ type DrinkCoffee struct{}
  * Movement action payload
  */
 type Move struct {
-	Speed float64
+	Speed float32
 }
 
 /*
